feat: implement rays and spheres
